models/requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
cryptography==42.0.5
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.0
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
fsspec==2024.3.1
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.20.2
huggingface-hub==0.22.2
idna==3.7
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pypdfium2==4.29.0
pytest==8.1.1
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
PyYAML==5.4.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
safetensors==0.4.2
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sympy==1.12
systemd-python==234
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
triton==2.2.0
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.1
wadllib==1.3.6
Werkzeug==3.0.2
zipp==1.0.0
